refactor: auth and user controllers are now in seperate files
