+++ backend/src/controller/auth.ts
@@ -98,89 +98,4 @@ export async function userRegistration(
   res.json({ success: [{ msg: "New user has been created!" }], token: token });
 }
 
-// @GET - /api/v1/users/
-// All users
-export async function users(req: express.Request, res: express.Response) {
-  const userRepository = getConnection().getRepository(User);
-  const users = await userRepository.find({
-    select: ["id", "name", "email", "phone", "role"],
-    relations: ["addresses"],
-  });
-
-  res.json({ data: users });
-}
-
-// @GET - /api/v1/user
-// Individual user
-export async function getUser(req: express.Request, res: express.Response) {
-  const id = req.params.userId;
-
-  const userRepository = getConnection().getRepository(User);
-  const findUserById = await userRepository.findOne(
-    { id },
-    {
-      relations: ["addresses"],
-    }
-  );
-
-  if (!findUserById) {
-    return res.status(400).json({ errors: [{ msg: "User not found" }] });
-  }
-
-  res.json({ msg: "User found", data: findUserById });
-}
-
-// @PUT /v1/api/users/:userId
-// update an user
-export async function updateUser(req: express.Request, res: express.Response) {
-  const userId = req.params.userId;
-  const { name, email, phone, role } = req.body;
-
-  const userRepository = getConnection().getRepository(User);
-  const userToUpdate = await userRepository.findOne({ id: userId });
-  if (userToUpdate) {
-    try {
-      const newUser = new User();
-      newUser.name = name;
-      newUser.email = email;
-      newUser.phone = phone;
-      newUser.role = role;
-      await userRepository.update(userId, newUser);
-      res.status(200).json({ success: [{ msg: "User info updated" }] });
-    } catch (e) {
-      res
-        .status(400)
-        .json({ errors: [{ msg: "User info could not be updated" }] });
-    }
-  } else {
-    res.status(400).json({
-      errors: [{ msg: "User not found or invalid url parameter(userId)" }],
-    });
-  }
-}
-
-// @DELETE /v1/ap/users/:userId
-// delete an user
-export async function deleteUser(req: express.Request, res: express.Response) {
-  const userId = req.params.userId;
-
-  const userRepository = getConnection().getRepository(User);
-  const userToDelete = await userRepository.findOne({ id: userId });
-
-  if (userToDelete) {
-    try {
-      await userRepository.delete(userId);
-      res.json({ data: "User deleted" });
-    } catch (e) {
-      res.status(400).json({ errors: [{ msg: "User could not be deleted" }] });
-    }
-  } else {
-    res.status(400).json({
-      errors: [
-        { msg: "User could not be deleted or invalid url parameter(userId)" },
-      ],
-    });
-  }
-}
-
 export default router;
